Move plugin from test to core
