+++ tests/codegen/test_print_operation.py
@@ -1,0 +1,23 @@
+from pathlib import Path
+
+import pytest
+
+from strawberry.codegen import QueryCodegen
+from strawberry.codegen.plugins.print_operation import PrintOperationPlugin
+
+
+HERE = Path(__file__).parent
+QUERIES = list(HERE.glob("queries/*.graphql"))
+
+
+@pytest.mark.parametrize("query", QUERIES, ids=[x.name for x in QUERIES])
+def test_codegen(
+    query: Path,
+    schema,
+):
+    generator = QueryCodegen(schema, plugins=[PrintOperationPlugin()])
+    query_content = query.read_text()
+
+    result = generator.codegen(query_content)
+
+    assert result == query_content
